requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
PyYAML==6.0.1
websockets==12.0
